refactor: 레포지토리와 서비스를 인터페이스 분리
