+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -8,22 +8,22 @@ import org.springframework.web.bind.annotation.RestController;
 import racingcar.controller.dto.RacingInfoRequest;
 import racingcar.controller.dto.RacingInfoResponse;
 import racingcar.domain.CarGroup;
-import racingcar.service.RacingGameService;
+import racingcar.service.RacingGameServiceImpl;
 
 @RestController
 public class RacingGameController {
 
-    private final RacingGameService racingGameService;
+    private final RacingGameServiceImpl racingGameServiceImpl;
 
-    public RacingGameController(final RacingGameService racingGameService) {
-        this.racingGameService = racingGameService;
+    public RacingGameController(final RacingGameServiceImpl racingGameServiceImpl) {
+        this.racingGameServiceImpl = racingGameServiceImpl;
     }
 
     @PostMapping("/plays")
     public ResponseEntity<RacingInfoResponse> playRacingGame(@RequestBody RacingInfoRequest request) {
         final String[] carNames = request.getNames().split(",");
         final CarGroup carGroup = new CarGroup(carNames);
-        final RacingInfoResponse response = racingGameService.race(carGroup, request.getCount());
+        final RacingInfoResponse response = racingGameServiceImpl.race(carGroup, request.getCount());
 
         return ResponseEntity.ok(response);
     }
