+++ src/main/java/racingcar/repository/RacingGameRepositoryImpl.java
@@ -1,0 +1,33 @@
+package racingcar.repository;
+
+import java.sql.PreparedStatement;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class RacingGameRepositoryImpl implements RacingGameRepository {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public RacingGameRepositoryImpl(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public int save(final String winners, final int count) {
+        final String sql = "insert into racing_game(winners, trial) values(?, ?)";
+        final KeyHolder keyHolder = new GeneratedKeyHolder();
+
+        jdbcTemplate.update(con -> {
+            final PreparedStatement ps = con.prepareStatement(sql, new String[]{"id"});
+            ps.setString(1, winners);
+            ps.setInt(2, count);
+            return ps;
+        }, keyHolder);
+
+        return keyHolder.getKey().intValue();
+    }
+}
