+++ src/test/java/racingcar/service/RacingGameServiceImplTest.java
@@ -15,10 +15,10 @@ import racingcar.controller.dto.RacingInfoResponse;
 import racingcar.domain.CarGroup;
 
 @SpringBootTest
-class RacingGameServiceTest {
+class RacingGameServiceImplTest {
 
     @Autowired
-    private RacingGameService service;
+    private RacingGameServiceImpl service;
 
     @DisplayName("자동차 경주를 통해 게임 결과를 반환한다.")
     @Test
